Initial refactor and imports checked
