+++ packages/cli/src/cli.js
@@ -1,8 +1,8 @@
 import { program } from "commander";
 import inquirer from "inquirer";
 import requireAll from "require-all";
-import { log } from "./core"
-import { init } from './scripts'
+import * as log  from "./core/log.js"
+import * as init  from './scripts/init.js'
 import 'dotenv/config';
 
 /**
