+++ packages/cli/src/core/config.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-useless-escape */
 const { NODE_ENV, DEBUG } = process.env;
-const appRoot = require("app-root-path");
+import appRoot from "app-root-path";
 
 const isDebug = DEBUG === "true";
 const isLocalEnv = NODE_ENV === "local";
