+++ packages/cli/src/core/index.js
@@ -1,6 +1,6 @@
-import CICDError from "./error";
-import log from "./log";
-import utils from "./utils";
-import common from "./common";
+import CICDError from "./error.js";
+import * as log  from "./log.js";
+import utils from "./utils.js";
+import common from "./common.js";
 
 export { CICDError, common, log, utils };
