+++ packages/cli/src/core/utils.js
@@ -1,4 +1,4 @@
-import log from "./log.mjs";
+import * as log  from "./log.js";
 import properties from "properties";
 import fs from "fs";
 import { workflowProps, PROPS_FILES_CONFIG } from "./config";
