+++ packages/cli/src/scripts/index.js
@@ -1,3 +1,3 @@
-import init from "./init";
+import * as init  from "./init.js";
 
 export { init };
